Update week13.md
## CLASS NOTES (12/03)

__Today__
- Work session
- Project check-ins

__HOMEWORK (DUE 12/10)__

__COMPLETE PROTOTYPE ALPHA__

Create a playable prototype of your game and upload to each team member's class Drive folder by the start of next class.

An alpha prototype is a functionally complete prototype of your game that we can test and give feedback on next week. It should be highly representative of the experience you want your players to have.

We will test each prototype and give feedback next week in our group meeting so please make sure to have builds available by the start of class.